move input override code that the renderer needs
